+++ packages/content-research/types/wordpress.d.ts
@@ -1,7 +1,0 @@
-interface ContentResearchData {
-	enabled: boolean;
-}
-
-interface Window {
-	contentResearchData?: ContentResearchData;
-}